handle index from oid value for metrics
