+++ pkg/collector/corechecks/snmp/checkconfig/config_validate_enrich_test.go
@@ -384,7 +384,7 @@ func Test_validateEnrichMetrics(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			errors := validateEnrichMetrics(tt.metrics)
+			errors := ValidateEnrichMetrics(tt.metrics)
 			assert.Equal(t, len(tt.expectedErrors), len(errors), fmt.Sprintf("ERRORS: %v", errors))
 			for i := range errors {
 				assert.Contains(t, errors[i], tt.expectedErrors[i])
