+++ pkg/collector/corechecks/snmp/report/report_metrics.go
@@ -76,7 +76,7 @@ func (ms *MetricSender) reportColumnMetrics(metricConfig checkconfig.MetricsConf
 		for fullIndex, value := range metricValues {
 			// cache row tags by fullIndex to avoid rebuilding it for every column rows
 			if _, ok := rowTagsCache[fullIndex]; !ok {
-				rowTagsCache[fullIndex] = append(common.CopyStrings(tags), metricConfig.MetricTags.GetTags(fullIndex, values)...)
+				rowTagsCache[fullIndex] = append(common.CopyStrings(tags), getTagsUsingMetricConfig(metricConfig.MetricTags, fullIndex, values)...)
 			}
 			rowTags := rowTagsCache[fullIndex]
 			ms.sendMetric(symbol.Name, value, rowTags, metricConfig.ForcedType, metricConfig.Options)
