+++ pkg/collector/corechecks/snmp/test/metadata_conf.d/snmp.d/profiles/_generic-if.yaml
@@ -42,6 +42,14 @@ metadata:
           OID: 1.3.6.1.2.1.31.1.1.1.1
           name: ifName
         tag: interface
+      - column:
+          OID: 1.3.6.1.2.1.4.20.1.1
+          name: ipAdEntAddr
+          index_from_oid_value:
+            OID: 1.3.6.1.2.1.4.20.1.2
+            name: ipAdEntIfIndex
+        table: ipAddrTable
+        tag: ip_address
 
 metrics:
 - MIB: IF-MIB
